requirements.txt
arrow==1.3.0
asgiref==3.7.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==4.0.0
bootstrap4==0.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
crispy-tailwind==1.0.0
cryptography==42.0.0
defusedxml==0.8.0rc2
Django==5.0.1
django-bootstrap4==23.4
django-bootstrap5==23.4
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-social-auth==0.7.28
django-social-share==2.3.0
django-summernote==0.8.20.0
django-taggit==5.0.1
django-tailwind==3.8.0
factory-boy==3.3.0
Faker==22.7.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.0.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-openid==2.2.5
python-slugify==8.0.1
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.1.0
webencodings==0.5.1
Whoosh==2.7.4
